Made last minute changes
