+++ src/main/java/command/AddDeadlineCommand.java
@@ -1,3 +1,6 @@
+package command;
+import storage.Storage;
+import task.Deadline;
 import task.TaskList;
 import ui.Ui;
