+++ src/main/java/command/AddEventCommand.java
@@ -1,3 +1,7 @@
+package command;
+
+import storage.Storage;
+import task.Event;
 import task.TaskList;
 import ui.Ui;
