+++ src/main/java/command/FindCommand.java
@@ -1,3 +1,6 @@
+package command;
+
+import storage.Storage;
 import task.Task;
 import task.TaskList;
 import ui.Ui;
